move the code in the pre-processing module
